// src/openapi/models/http-log-entry-details-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * TRON Common API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface HttpLogEntryDetailsDto
 */
export interface HttpLogEntryDetailsDto {
    /**
     * 
     * @type {string}
     * @memberof HttpLogEntryDetailsDto
     */
    requestTimestamp?: string;
    /**
     * 
     * @type {string}
     * @memberof HttpLogEntryDetailsDto
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof HttpLogEntryDetailsDto
     */
    requestedUrl?: string;
    /**
     * 
     * @type {string}
     * @memberof HttpLogEntryDetailsDto
     */
    requestBody?: string;
    /**
     * 
     * @type {string}
     * @memberof HttpLogEntryDetailsDto
     */
    remoteIp?: string;
    /**
     * 
     * @type {string}
     * @memberof HttpLogEntryDetailsDto
     */
    requestMethod?: string;
    /**
     * 
     * @type {string}
     * @memberof HttpLogEntryDetailsDto
     */
    responseBody?: string;
    /**
     * 
     * @type {string}
     * @memberof HttpLogEntryDetailsDto
     */
    requestHost?: string;
    /**
     * 
     * @type {string}
     * @memberof HttpLogEntryDetailsDto
     */
    userAgent?: string;
    /**
     * 
     * @type {string}
     * @memberof HttpLogEntryDetailsDto
     */
    queryString?: string;
    /**
     * 
     * @type {number}
     * @memberof HttpLogEntryDetailsDto
     */
    timeTakenMs?: number;
    /**
     * 
     * @type {string}
     * @memberof HttpLogEntryDetailsDto
     */
    userName?: string;
    /**
     * 
     * @type {number}
     * @memberof HttpLogEntryDetailsDto
     */
    statusCode?: number;
}